Update Bylaws_of_The_Bitcoin_Foundation.md
Section 5.3 (b) Per the updated industry member benefits, we need to break out the three industry seats so that:
- the seat currently occupied by Mark is voted on only by Platinum members
- the seat currently occupied by Charlie is voted on only by Gold and Silver members
- the third seat up for election in September 2013 is voted on by all industry members

We also need to update Schedule A to reflect current membership pricing. Best way might be to list the USD values that are constant, with a note that they are paid in bitcoin based on market price.