Reduce quorum form ⅔ to ½ for director elections
SITUATION: The most recent election was not definitive in one round due
to failure to achieve approval by a majority of a quorum.

BACKGROUND: The current quorum for director elections is two-thirds of
the corresponding member class. An election winner must achieve a
majority of this quorum. Quorum serves two purposes: 1) establishes
that a sufficient fraction of the membership have been reached with a
ballot and 2) a sufficient fraction of the members wish the ballot to
be brought for a vote.

However, in the most recent election we had several voters who wished
for the election to be held but did not have a preference for any
candidate. While ballots not supporting a particular candidate do help
achieve quorum, such ballots also raise the bar that a successful
candidate must achieve. This is not the intent of quorum.

ACTION: This is the first of two pull requests. In this request, the
quorum requirement is lowered to one-half from two-thirds. A subsequent
pull request will describe a method for reducing the "majority of a
quorum" requirement to reflect a majority of those that submitted a
ballot supporting a candidate after quorum requirement achieved.

RESPONSE: Pull request submitted for community and board discussion.