Require Majority of "Voting" Voters
Pull request #2 in series. See
https://github.com/pmlaw/The-Bitcoin-Foundation-Legal-Repo/pull/20/

The aim of this pull is to set the bar for electing a director to a
majority of the voters who cast a ballot in favor of a candidate.

This would permit members to participate in quorum without supporting a
specific candidate and without raising the number of required approvals
to elect a director.