Update Article II (Purpose): Anonymity and Funding
This pull request is responsive to issue #19 ("Anonymity and Funding").  It is expressly written to minimize change to the Bylaws, provides an additional emphasis on anonymity that was not previously present, while still ensuring that bitcoin in its current, non-anonymous form is expressly supported in the Purpose section.  Where the Corporation "shall promote and protect" aspects of bitcoin use (and use of "such systems,") "all derivative uses of blockchain data" are added to emphasize that the Corporation shall promote and protect the many uses of bitcoin.  Additionally, a reference to "financial privacy" is replaced with "user-defined privacy," emphasizing the vital role of individual choice which is present in the Bylaws.