Change Quorum to "Active" Members for Board Elections
TOPIC: Change requirements for electing board members.

SCOPE: Restricted to elections of board members.

ISSUE:
1) Inactive members make achieving quorum difficult for electing board
members.

PROPOSED CHANGES:
1) Reduce quorum fraction from 2/3 to 1/2
2) Quorum will be established as a fraction of the “active” members
	a) Define “active” members as those who confirm themselves as active

COMMENTARY: As the organization matures, members will invariably become
inactive, especially lifetime members.

In the current state of the bylaws, inactive members essentially
function as members who wish all matters submitted to membership vote
to fail to achieve quorum and no candidate to be successful. While such
wishes must be a choice each member is free to choose, it should not be
the default state assigned if no action by a member is taken.

The definition of “active” membership prevents inactive members from
increasing quorum requirements and instead makes inactive members of no
impact to the business of electing directors.

UNRESOLVED ISSUES:
1) Withheld ballots help for establishing quorum but increase the
number of approvals required to elect a director. Thus, members who
wish to facilitate foundation business are effectively required to
support a candidate. Decoupling quorum from required approvals would
permit members more freedom to facilitate foundation elections without
having to support a particular candidate.

POSSIBLE FUTURE CHANGES NOT INCLUDED IN THIS PULL REQUEST:
1) Set number of required approval to a majority of ballots supporting
a candidate (e.g., “ the number of approvals required to elect a
director shall be the majority of those votes cast that vote for a
particular candidate in that election.”).

REASON FOR NON-INCLUSION:
The cryptographic voting system intended for use in this election does
not support internal ballot logic (example, if “withheld” option
selected on ballot, do not permit candidates to be selected). Thus,
approval voting with ballots incapable of implementing logic could
permit voters to amplify their voice by both voting for candidates and
selecting withheld.

It may be feasible to modify the voting software to support ballots
that can implement logic. This may be more difficult than anticipated
given that auditing an election would require not only a proof that the
ballots tallied correctly but also that no ballot was counted that
violated the embedded logic.